+++ Tycoonia/Domain/Buildings/Factory/FactoryConcrete.cs
@@ -11,7 +11,7 @@ namespace Tycoonia.Domain.Buildings.Factory
                 { "Money", 100 },
                 { "Bricks", 1 }
             };
-            ProductionRate = 1;
+            ProductionRate = 1m;
             EnergyConsumption = 0.5m;
             RecipeList = new Dictionary<string, byte>
             {
@@ -20,7 +20,7 @@ namespace Tycoonia.Domain.Buildings.Factory
                 { "Gravel", 2 },
                 { "Water", 1 }
             };
-            ProductionItemList = new Dictionary<string, int>
+            ProductionItemList = new Dictionary<string, decimal>
             {
                 { "Concrete", ProductionRate }
             };
