+++ Tycoonia/Domain/Buildings/Factory/FactoryEnrichmentUranium.cs
@@ -12,17 +12,17 @@
                 { "Money", 100 },
                 { "Bricks", 1 }
             };
-            ProductionRate = 1;
+            ProductionRate = 1m;
             EnergyConsumption = 0.5m;
             RecipeList = new Dictionary<string, byte>
             {
                 { "Money", 10 },
                 { "Uranium", 3 }
             };
-            ProductionItemList = new Dictionary<string, int>
+            ProductionItemList = new Dictionary<string, decimal>
             {
                 { "Uranium-238", ProductionRate},
-                { "Uranium-235", 1}
+                { "Uranium-235", 1m}
             };
         }
     }
