+++ Tycoonia/Domain/Buildings/Factory/FactoryThoriumRod.cs
@@ -13,7 +13,7 @@ namespace Tycoonia.Domain.Buildings.Factory
                 { "Money", 100 },
                 { "Bricks", 1 }
             };
-            ProductionRate = 1;
+            ProductionRate = 1m;
             EnergyConsumption = 0.5m;
             RecipeList = new Dictionary<string, byte>
             {
@@ -21,7 +21,7 @@ namespace Tycoonia.Domain.Buildings.Factory
                 { "Thorium-232", 1},
                 { "Plutonium-239", 1 }
             };
-            ProductionItemList = new Dictionary<string, int>
+            ProductionItemList = new Dictionary<string, decimal>
             {
                 { "Thorium Rod", ProductionRate }
             };
