+++ Tycoonia/Domain/Buildings/Factory/FactoryUraniumRod.cs
@@ -11,7 +11,7 @@
                 { "Money", 100 },
                 { "Bricks", 1 }
             };
-            ProductionRate = 1;
+            ProductionRate = 1m;
             EnergyConsumption = 0.5m;
             RecipeList = new Dictionary<string, byte>
             {
@@ -19,7 +19,7 @@
                 { "Uranium-238", 1},
                 { "Uranium-235", 10}
             };
-            ProductionItemList = new Dictionary<string, int>
+            ProductionItemList = new Dictionary<string, decimal>
             {
                 { "Uranium Rod", ProductionRate }
             };
